Improve logic for waiting for server
The previous test was not differentiating an error resulting from lack
of a socket and an error resulting from a timeout. Since there may not
be a socket created for the server when we start waiting, we need to
keep trying up to our time limit even if the socket doesn't exist yet.