Rework `Source` data structure
`Source` was previously an abstraction letting us load the content of
files into strings. This change makes it instead a representation of the
input sources that Apalache needs in order to parse, typecheck, or model
check specs.

The bulk of the complexity here comes from functions used to find the names
of module dependencies from within the text of a TLA file.