Move load_input into trans_explorer
This is atavistic, and is only really needed for support for the WIP
state of the `trans_explorer`. This should be replaced by use of a
`Source` class instead, but work on that will left for a followup, as it
also calls for changes on the Scala RPC side.

Tracked in  #33